requirements.txt
tensorflow
keras
pandas
numpy
jupyter
notebook
tqdm
opencv-contrib-python
scikit-learn
streamlit
pillow
flask
